backstage/flask_demo/requirements.txt
Click==7.0
Flask-Cors==3.0.7
Flask-RESTful==0.3.6
Flask==1.0.2
Jinja2==2.10
MarkupSafe==1.1.0
PyJWT==1.7.1
Werkzeug==0.14.1
aliyun-python-sdk-core-v3==2.11.2
aliyun-python-sdk-kms==2.5.0
amqp==2.3.2
aniso8601==4.0.1
atomicwrites==1.2.1
attrs==18.2.0
billiard==3.5.0.5
celery==4.2.1
certifi==2018.11.29
chardet==3.0.4
crcmod==1.7
gevent==1.3.7
greenlet==0.4.15
gunicorn==19.9.0
idna==2.8
itsdangerous==1.1.0
jieba==0.39
kombu==4.2.2
marshmallow==2.16.3
mongoengine==0.16.3
more-itertools==4.3.0
oss2==2.6.0
pluggy==0.8.0
py==1.7.0
pycryptodome==3.7.2
pymongo==3.7.2
pytest==4.0.2
pytz==2018.7
redis==3.0.1
requests==2.21.0
six==1.12.0
urllib3==1.24.1
vine==1.1.4